task is completed:)

+++ app/src/main/java/com/example/testtask/model/database/DatabaseComment.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "comments_table")
 data class DatabaseComment(
-    @PrimaryKey val commentId: Int,
-    @ColumnInfo val postId: Int,
+    @PrimaryKey val commentId: Long,
+    @ColumnInfo val postId: Long,
     @ColumnInfo val name: String,
     @ColumnInfo val body: String,
     @ColumnInfo val email: String
