+++ app/src/main/java/com/example/testtask/model/database/DatabasePost.kt
@@ -4,8 +4,8 @@ import androidx.room.*
 
 @Entity(tableName = "posts_table")
 data class DatabasePost(
-    @PrimaryKey val postId: Int,
-    @ColumnInfo val userId: Int,
+    @PrimaryKey val postId: Long,
+    @ColumnInfo val userId: Long,
     @ColumnInfo val title: String,
     @ColumnInfo val body: String
 )
